Ensure the package-lock.json file gets committed too
If it doesn't the dependencies update in package.json, but package-lock.json gets left behind so clean-install ends up failing as a result. That said, npm install may fail on version conflicts, so the safest thing to do is to let it run, but if it fails, revert package-lock.json back to its original state and create the pull request as usual.