Add scripts to switch between local and live versions of components package
It's often useful when developing components to be able to test it out in the actual repository we're working in before it gets published. The most common way is probably to to do npm install ../components, but that doesn't often work well since the dependencies don't really resolve properly, meaning that the theme often looks broken when developing locally.

To get around this, we now generate a .tgz file using npm pack, which is what actually gets used when the package gets published, so we can just install that instead. To help make this easier to develop with, I have set up a few scripts in package.json to help us easily switch between the two without too much thought.