Deal with dependency warning on installation
react-code-blocks kept causing that issue in the terminal where it kept recommending we choose a different dependency. As such, I have chosen a different dependency - react-code-block (singular). It's a little harder to use as it requires more manual work, but I can still wrap around it as needed to ensure consistent behaviour.